XC-220: move initial_state to test fixtures
