+++ rs/ethereum/cketh/minter/src/lifecycle/tests.rs
@@ -2,6 +2,7 @@ mod init {
     use crate::lifecycle::init::InitArg;
     use crate::numeric::{TransactionNonce, Wei};
     use crate::state::{InvalidStateError, State};
+    use crate::test_fixtures::valid_init_arg;
     use assert_matches::assert_matches;
     use candid::{Nat, Principal};
     use num_bigint::BigUint;
@@ -89,19 +90,5 @@ mod init {
             state.eth_transactions.next_transaction_nonce(),
             TransactionNonce::ZERO
         );
-    }
-
-    fn valid_init_arg() -> InitArg {
-        InitArg {
-            ethereum_network: Default::default(),
-            ecdsa_key_name: "test_key_1".to_string(),
-            ethereum_contract_address: None,
-            ledger_id: Principal::from_text("apia6-jaaaa-aaaar-qabma-cai")
-                .expect("BUG: invalid principal"),
-            ethereum_block_height: Default::default(),
-            minimum_withdrawal_amount: Nat::from(10_000_000_000_000_000_u64),
-            next_transaction_nonce: TransactionNonce::ZERO.into(),
-            last_scraped_block_number: Default::default(),
-        }
     }
 }
